2018-08-12: By Ade

+++ resources/views/sales/detail.blade.php
@@ -11,7 +11,6 @@ Detail Sales
     <small>Preview</small>
   </h1>
   <ol class="breadcrumb">
-    <li><a href="/"><i class="fa fa-dashboard"></i> Dashboard</a></li>
     <li><a href="/master/sales">Sales</a></li>
     <li class="active">Detail Sales</li>
   </ol>
@@ -19,10 +18,24 @@ Detail Sales
 
 <section class="content">
 
+  @if(Session::has('success'))
+  <div class="alert alert-success alert-dismissible">
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+    <h4><i class="icon fa fa-check"></i> Success!</h4>
+    {{Session::get('success')}}.
+  </div>
+  @elseif(Session::has('warning'))
+  <div class="alert alert-warning alert-dismissible">
+    <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+    <h4><i class="icon fa fa-warning"></i> Warning!</h4>
+    {{Session::get('warning')}}.
+  </div>
+  @endif
+
   <div class="box box-default">
     <div class="box-header with-border">
       <a href="/master/sales/ubah/{{$sales->id}}" class="btn btn-sm btn-warning"><span class="fa fa-edit"></span></a>
-      <a href="/master/sales/hapus/{{$sales->id}}" class="btn btn-sm btn-danger"><span class="fa fa-trash"></span></a>
+      <a onclick="return confirm('Anda yakin ?')" href="/master/sales/hapus/{{$sales->id}}" class="btn btn-sm btn-danger"><span class="fa fa-trash"></span></a>
 
       <div class="box-tools pull-right">
         <button type="button" class="btn btn-box-tool" data-widget="collapse"><i class="fa fa-minus"></i></button>
@@ -34,7 +47,18 @@ Detail Sales
         <tr>
           <td width="15%"><strong>ID Sales</strong></td>
           <td width="1%">:</td>
-          <td>{{$sales->id}}</td>
+          <td width="30%">{{$sales->id}}</td>
+          <td width="1%" rowspan="6"><b>Foto</b></td>
+          <td width="1%" rowspan="6"><b>:</b></td>
+          <td rowspan="6">
+            <center>
+              @if ($sales->image == null || $sales->image=="")
+                <img src="{{asset('images/sales/null.png')}}" height="200px" alt="">
+              @else
+                <img src="{{asset('images/sales/'.$sales->id.'/'.$sales->image)}}" height="200px" alt="">
+              @endif
+            </center>
+          </td>
         </tr>
         <tr>
           <td><strong>Nama</strong></td>
